Add initial version of filesystem
Added basic in-memory implementation of filesystem using a Trie.
Implemented basic operations on files and directories.
Added basic unit-tests.

The filesystem is thread-safe and supports a clean interface for
testing/exposing it to command-line interface, http-server...etc.

A trie is used because it's efficient in terms of traversals both
up and down, supports prefix matching efficiently (i.e., tab completion),
and is also memory efficient if there are a ton of files that share
prefixes.